Added work around to be able to run actions with files from CLI.
When an action is run from CLI there are no mappings, and because of
that no schema can be resolved.
When files are used as parameters there is a check to see if a file
server is configured. This check requires current service schema.
The work around handles this exception.

Resolves kusanagi/katana#79.